fix: remove trailing slash from url
